Move couch purge service to libs
