+++ test/libs/couch/purge.spec.ts
@@ -4,7 +4,7 @@ import { expect } from 'chai';
 import sinon, { SinonStub } from 'sinon';
 import { ChtClientService } from '../../../src/services/cht-client';
 import { HttpClientRequest } from '@effect/platform';
-import { purgeFrom } from '../../../src/services/couch/purge';
+import { purgeFrom } from '../../../src/libs/couch/purge';
 import { genWithLayer, sandbox } from '../../utils/base';
 
 const FAKE_CLIENT_REQUEST = { hello: 'world' } as const;
@@ -17,7 +17,7 @@ const run = Layer
   .succeed(ChtClientService, { request: couchRequest } as unknown as ChtClientService)
   .pipe(genWithLayer);
 
-describe('Couch Purge Service', () => {
+describe('Couch Purge libs', () => {
   let requestSchemaBody: SinonStub;
   let requestPost: SinonStub;
