+++ test/services/purge.spec.ts
@@ -5,7 +5,7 @@ import * as pouchDBService from '../../src/services/pouchdb';
 import { PouchDBService } from '../../src/services/pouchdb';
 import { expect } from 'chai';
 import { PurgeService } from '../../src/services/purge';
-import * as couchPurgeService from '../../src/services/couch/purge';
+import * as couchPurgeService from '../../src/libs/couch/purge';
 import { genWithLayer, sandbox } from '../utils/base';
 import { ChtClientService } from '../../src/services/cht-client';
